move interpolate to ranking.py
